cleanup-requirements.txt
# Requirements for cleanup scripts
# Install with: pip install -r cleanup-requirements.txt

boto3>=1.28.0
python-dotenv>=1.0.0
requests>=2.31.0
